Create own component for annotation form additional texts section
